+++ src/main/java/kr/codesquad/ladder/domain/User.java
@@ -1,4 +1,4 @@
-package kr.codesquad.domain;
+package kr.codesquad.ladder.domain;
 
 public class User {
     private String name;
