+++ src/main/java/kr/codesquad/ladder/util/RandomBoolean.java
@@ -1,4 +1,4 @@
-package kr.codesquad.util;
+package kr.codesquad.ladder.util;
 
 import java.util.ArrayList;
 import java.util.List;
