+++ src/main/java/kr/codesquad/ladder/util/RandomUtil.java
@@ -1,4 +1,4 @@
-package kr.codesquad.util;
+package kr.codesquad.ladder.util;
 
 import java.util.List;
