+++ src/main/java/kr/codesquad/ladder/view/InputView.java
@@ -1,4 +1,4 @@
-package kr.codesquad.view;
+package kr.codesquad.ladder.view;
 
 import java.io.BufferedReader;
 import java.io.IOException;
